Validation added
Api of post count with user data
reponse set

+++ app/Http/Controllers/API/CommentController.php
@@ -16,8 +16,24 @@ use App\Models\EventCategory;
 use Illuminate\Support\Facades\Auth;
 
 class CommentController extends Controller {
+    
     public function store(Request $request, $postId) {
-        $request->validate(['body' => 'required|string', 'emojis' => 'nullable|array']);
+        // $request->validate(['body' => 'required|string', 'emojis' => 'nullable|array']);
+
+        //manual validation due to sendError custom method for error responses
+        $validator = \Validator::make($request->all(), [
+            'body' => 'required|string',
+            'emojis' => 'nullable|array'
+        ]);
+        if ($validator->fails()) {
+            return $this->sendError('Validation Error', $validator->errors(), 422);
+        }
+
+        //post check and validation
+        if (!Post::where('id', $postId)->exists()) {
+            return $this->sendError('Post not found', [], 404);
+        }
+      
         $comment = Comment::create([
             'post_id' => $postId,
             'user_id' => auth()->id(),
@@ -34,6 +50,12 @@ class CommentController extends Controller {
     //     return response()->json(['message' => 'Comment liked']);
     // }
     public function like($id) {
+        // Check if the comment exists
+        // If not, return an error response
+        if (!Comment::where('id', $id)->exists()) {
+            return $this->sendError('Comment not found', [], 404);
+        }
+     
         $like = CommentLike::where('comment_id', $id)->where('user_id', auth()->id())->first();
         if ($like) {
             $like->delete();
@@ -49,7 +71,27 @@ class CommentController extends Controller {
     }
 
     public function reply(Request $request, $id) {
-        $request->validate(['body' => 'required|string', 'emojis' => 'nullable|array']);
+
+        // $request->validate(['body' => 'required|string', 'emojis' => 'nullable|array']);
+
+
+        // Check if the comment exists
+        // If not, return an error response
+        if (!Comment::where('id', $id)->exists()) {
+            return $this->sendError('Comment not found', [], 404);
+        }
+
+        //manual validation due to sendError custom method for error responses
+
+        $validator = \Validator::make($request->all(), [
+            'body' => 'required|string',
+            'emojis' => 'nullable|array'
+        ]);
+        if($validator->fails()){
+            return $this->sendError('Validation Error', $validator->errors(),422);
+        }
+
+
         $reply = Reply::create([
             'comment_id' => $id,
             'user_id' => auth()->id(),
